Throttle various NVAPI HW sensor readings back even farther and stop the collection of sensor data altogehter when the GPU widget or text OSD is not visible, because NV royally messed up the performance of HW sensors on 50 series GPUs.

+++ include/SpecialK/performance/gpu_monitor.h
@@ -130,8 +130,8 @@ struct gpu_sensors_t
     bool     queried_nv_pstates = false;
 
     struct {
-      int phase0 = 0;
-      int phase1 = 0;
+      int64 phase0 = 0;
+      int64 phase1 = 0;
     } amortization;
   } gpus [MAX_GPUS];
